+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 ### Added
 * Tour guide for new features (`OSIDB-4583`)
 * Allow grouping affects by ps_module (`OSIDB-4546`)
+* Support revert and reset workflow functions (`OSIDB-4489`)
 
 
 ## [2025.10.0]